+++ src/util/mod.rs
@@ -4,4 +4,4 @@ pub mod datetime;
 pub mod vector;
 pub mod hash;
 pub mod country;
-pub mod csv;
+pub mod spotify_csv;
